requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloud-sql-python-connector==1.15.0
cryptography==44.0.0
dnspython==2.7.0
Flask==3.1.0
frozenlist==1.5.0
gcloud==0.18.3
google-auth==2.37.0
googleapis-common-protos==1.66.0
greenlet==3.1.1
httplib2==0.22.0
idna==3.10
importlib_metadata==8.5.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
multidict==6.1.0
mysql-connector-python==9.1.0
oauth2client==4.1.3
propcache==0.2.1
protobuf==5.29.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.2.0
requests==2.32.3
rsa==4.9
six==1.17.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
yarl==1.18.3
zipp==3.21.0
